LPS-158854 Move EntityExtensionThreadLocal to api
